created cloud_project class, moved shared resources there
