+++ cloud_classifier/cloud_classifier.py
@@ -9,6 +9,8 @@ import tools.file_handling as fh
 import tools.confusion as conf
 import tools.write_netcdf as ncdf
 
+import cloud_project
+
 
 import importlib
 
@@ -21,81 +23,11 @@ importlib.reload(ncdf)
 importlib.reload(conf)
 
 
-class cloud_classifier:
+class cloud_classifier(cloud_project.cloud_project):
     def __init__(self, project_path=None):
 
-        self.project_path = project_path
-        self.param_handler = parameter_handler.parameter_handler()
-
-        self.params = self.param_handler.parameters
-        self.filelists = self.param_handler.filelists
-
-        self.masked_indices = None
+        super().__init__(project_path)
         self.trainer = cloud_trainer.cloud_trainer()
-
-    # ############ CREATING, LOADING AND SAVING PROJECTS ######################
-    # ########################################################################
-
-    def create_new_project(self, name, path=None):
-        """
-        Creates a persistant classifier project.
-
-
-        Parameters
-        ----------
-        name : string
-            Name of the the project that will be created
-
-        path : string, optional
-            Path to the directory where the project will be stored. If none is
-            given, the current working directory will be used.
-        """
-
-        if path is None:
-            path = os.getcwd()
-
-        folder = os.path.join(path, name)
-        if os.path.isdir(folder):
-            print("Folder with given name already exits! Loading existing project!")
-        else:
-            try:
-                self.param_handler.initalize_settings(folder)
-                print("Project folder created successfully!")
-
-            except Exception:
-                raise RuntimeError(
-                    "Could not initalize project settings at given location!"
-                )
-        self.load_project(folder)
-
-    def load_project(self, path):
-        """
-        Loads a persistant classifier project.
-
-        Parameters
-        ----------
-        path : string
-            Path to the stored project
-        """
-        self.project_path = path
-        self.load_project_data()
-
-    def load_project_data(self):
-        if self.project_path is None:
-            raise ValueError("Project path not set")
-        self.param_handler.load_parameters(self.project_path)
-        self.param_handler.load_filelists(self.project_path)
-
-    def save_project_data(self):
-        if self.project_path is None:
-            raise ValueError("Project path not set")
-        self.param_handler.save_parameters(self.project_path)
-        self.param_handler.save_filelists(self.project_path)
-
-    def set_project_parameters(self, **kwargs):
-        self.param_handler.set_parameters(**kwargs)
-        self.param_handler.set_filelists(**kwargs)
-        self.save_project_data()
 
     ######################    PIPELINE  ######################################
     ##########################################################################
